reimagined/requirements.txt
# Core Django and web framework
Django>=4.2,<5.0
python-dotenv>=1.0.0

# API and HTTP clients
requests>=2.31.0

# Testing framework
pytest>=7.0.0
pytest-django>=4.5.0

# Optional: Development tools (uncomment as needed)
# ruff>=0.1.0
# black>=23.0.0
# isort>=5.12.0